requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
requests==2.31.0
python-dotenv==1.0.1
openai==1.35.10
groq==0.9.0
pytest==8.2.2
pypdf==4.3.1
httpx==0.27.0
